+++ src/db_adapter/BUILD
@@ -17,6 +17,7 @@ cc_library(
         ":metta_file_writer",
         "//commons:commons_lib",
         "//commons/atoms:atoms_lib",
+        "//db_adapter/mork:mork_lib",
         "//db_adapter/postgres:postgres_lib",
     ],
 )
