+++ src/db_adapter/mork/BUILD
@@ -6,6 +6,7 @@ cc_library(
     name = "mork_lib",
     includes = ["."],
     deps = [
+        ":mork_connection",
         ":mork_wrapper",
     ],
 )
@@ -15,17 +16,23 @@ cc_library(
     srcs = ["MorkWrapper.cc"],
     hdrs = ["MorkWrapper.h"],
     includes = ["."],
-    linkopts = [
-        "-L/usr/local/lib",
-        "-lpqxx",
-        "-lpq",
+    deps = [
+        ":mork_connection",
+        "//commons:commons_lib",
+        "//db_adapter:bounded_shared_queue",
+        "//db_adapter:database_types",
+        "//db_adapter:database_wrapper",
     ],
+)
+
+cc_library(
+    name = "mork_connection",
+    srcs = ["MorkConnection.cc"],
+    hdrs = ["MorkConnection.h"],
+    includes = ["."],
     deps = [
         "//atomdb/morkdb:morkdb_lib",
         "//commons:commons_lib",
-        "//db_adapter:bounded_shared_queue",
         "//db_adapter:database_connection",
-        "//db_adapter:database_mapper",
-        "//db_adapter:database_wrapper",
     ],
 )
