+++ src/db_adapter/mork/MorkWrapper.cc
@@ -1,42 +1,13 @@
 #include "MorkWrapper.h"
 
+#include "MorkConnection.h"
 #include "Utils.h"
 
 #define LOG_LEVEL INFO_LEVEL
 #include "Logger.h"
 
 using namespace std;
-using namespace atoms;
-
-// ==============================
-//  Construction / destruction
-// ==============================
-
-MorkConnection::MorkConnection(const string& id, const string& host, int port)
-    : DatabaseConnection(id, host, port) {
-    RAISE_ERROR("MorkConnection constructor not implemented yet");
-}
-
-MorkConnection::~MorkConnection() {}
-
-// ==============================
-// Public
-// ==============================
-
-void MorkConnection::connect() {
-    RAISE_ERROR("MorkConnection::connect() not implemented yet");
-    return;
-}
-
-void MorkConnection::disconnect() {
-    RAISE_ERROR("MorkConnection::disconnect() not implemented yet");
-    return;
-}
-
-vector<string> MorkConnection::query(const string& metta_query) {
-    RAISE_ERROR("MorkConnection::query() not implemented yet");
-    return {};
-}
+using namespace db_adapter;
 
 // ==============================
 //  Construction / destruction
