+++ src/db_adapter/mork/MorkWrapper.h
@@ -2,36 +2,21 @@
 
 #include <memory>
 #include <string>
-#include <vector>
 
 #include "BoundedSharedQueue.h"
+#include "DatabaseTypes.h"
 #include "DatabaseWrapper.h"
-#include "MorkDB.h"
+#include "MorkConnection.h"
 
 using namespace std;
-using namespace atomdb;
 
 namespace db_adapter {
-
-class MorkConnection : public DatabaseConnection {
-   public:
-    MorkConnection(const string& id, const string& host, int port);
-    ~MorkConnection();
-
-    void connect() override;
-    void disconnect() override;
-
-    vector<string> query(const string& metta_query);
-
-   protected:
-    unique_ptr<MorkClient> conn;
-};
 
 class MorkWrapper : public DatabaseWrapper {
    public:
     MorkWrapper(MorkConnection& conn,
                 shared_ptr<BoundedSharedQueue> output_queue,
-                MAPPER_TYPE mapper_type = MAPPER_TYPE::MORK2ATOMS);
+                MAPPER_TYPE mapper_type = MAPPER_TYPE::METTA2ATOMS);
     ~MorkWrapper();
 
     void map(const string& metta_query);
