+++ src/db_adapter/postgres/PostgresWrapper.cc
@@ -14,6 +14,7 @@
 #include "Logger.h"
 
 using namespace std;
+using namespace db_adapter;
 
 // ==============================
 //  Construction / destruction
